Remove pprint from setup.py
As discussed in this [SO post](https://stackoverflow.com/questions/63457762/error-could-not-find-a-version-that-satisfies-the-requirement-pprint-from-r-r), `pprint` is a standard library built-in module and therefore can't be listed as a requirement. The speculative reason why this previously worked is because there may have been a package on pypi named 'pprint' that was recently removed.

Note that this currently breaks `pip install bpnet`.